requirements.txt
click==7.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dotenv==0.12.0
redis==3.4.1
SQLAlchemy==1.3.16
Werkzeug==1.0.1
WTForms==2.2.1
WTForms-SQLAlchemy==0.1
